Fix failing test cases.
