+++ package.json
@@ -26,8 +26,8 @@
   "dependencies": {
     "@emotion/css": "^11.1.3",
     "@emotion/react": "^11.1.5",
-    "csstips": "^1.2.0",
     "color": "^3.1.3",
+    "csstips": "^1.2.0",
     "debounce": "^1.2.0",
     "rc-slider": "^9.6.2",
     "rc-tooltip": "^4.2.1",
@@ -50,14 +50,14 @@
     "@storybook/react": "^6.1.11",
     "@testing-library/jest-dom": "^5.11.6",
     "@testing-library/react": "^11.2.3",
+    "@types/color": "^3.0.1",
     "@types/debounce": "^1.2.0",
     "@types/jest": "^26.0.10",
     "@types/rc-tooltip": "^3.7.3",
-    "@types/react": "^16.9.48",
-    "@types/react-dom": "^16.9.8",
+    "@types/react": "^17.0.6",
+    "@types/react-dom": "^16.0.5",
     "@types/react-icons": "^3.0.0",
     "@types/react-router-dom": "^5.1.5",
-    "@types/color": "^3.0.1",
     "@typescript-eslint/eslint-plugin": "^3.5.0",
     "@typescript-eslint/parser": "^3.5.0",
     "babel-loader": "^8.1.0",
@@ -70,9 +70,9 @@
     "jest": "^26.6.3",
     "lint-staged": "^10.2.13",
     "prettier": "^2.0.5",
-    "react": "^16.8.6",
+    "react": "^17.0.2",
     "react-docgen-typescript-loader": "^3.7.2",
-    "react-dom": "^16.8.6",
+    "react-dom": "^17.0.2",
     "react-icons": "^4.2.0",
     "source-map-loader": "^1.0.2",
     "stylelint": "^13.6.1",
