+++ src/typings/surface.ts
@@ -1,0 +1,9 @@
+import { MarginClassNameInput } from '../helpers/margin';
+import { PaddingClassNameInput } from '../helpers/padding';
+
+export interface SurfaceProps {
+  elevation?: number;
+  border?: number;
+  padding?: PaddingClassNameInput;
+  margin?: MarginClassNameInput;
+}
